Add a socket cache
When a socket is closed, typically because a proxy is deleted,
don't close the file descriptor. Rather keep a cache which can
be looked-up and used when a subsequent proxy is started.

The aim is to avoid unbinding and binding to ports and thus
interrupting services unnecessarily. This will be used
in conjunction with a facility to allow haproxy to reinitialise
itself.