Add master/worker model
A master process is detached from the terminal and then spawns worker
processes. The master does nothing but waits for a signal on which it will
kill its workers.

A later patch will allow for restarting the workers.  To facilitate this
the master keeps all bound sockets open and by using the socket_cache these
can be re-used. This will enable haproxy to be restarted without refusing or
resetting connections to existing sockets.