Log using send_log from Alert and Warning
Alert and Warning are used in initialisation code, often
without any call to send_log. This assumes that stderr is
available which is currently the case. But when haproxy is
extended to be able to reinitialise it self such logging
may take place after haproxy is daemonised and thus detached
from the stderr. In order to make messages visible in
such situations Alert and Warning have been extended to also
log to send_log if possible.