Run master outside of chroot
On reset this allows the master process to behave in a way much closer
to the way that haproxy behaves when initially started. In particular
the configuration file can be accessed without needing to be copied
into the chroot.

Entering the chroot is unchanged when master/worker mode is
not enabled to allow any errors to be recoverable from
a soft-restart point of view. Soft-restart is not currently
implemented for master/worker mode.

Signed-off-by: Simon Horman <horms@verge.net.au>