dynamic health check
Support a dynamic health check performed by opening a TCP socket to a
pre-defined port and reading an ascii string. The string should have one of
the following forms:

i. An ascii representation of an positive integer percentage.
   e.g. "75%"

   Values in this format will set the wight proportional to the initial
   weight of a server as configured when haproxy starts.

ii. The string "drain".

   This will cause the weight of a server to be set to 0, and thus it will
   not accept any new connections other than those that are accepted via
   persistence.

ii. The string "disable".

   Put the server into maintenance mode. The server must be re-enabled
   before any further health checks will be performed.

A dynmaic helath check may be configued using "option dynamic-chk".
The use of an alternate check-port, used to obtain dynamic heath check
information described above as opposed to the port of the service,
may be useful in conjunction with this option.

e.g.

    option  dynamic-chk
    server  http1_1 10.0.0.10:80 check port 10000 weight 100

Signed-off-by: Simon Horman <horms@verge.net.au>