MEDIUM: checks: Add supplementary agent checks
Allow an agent check to be run in conjunction with one other server
health check.

If the backend for a server check is not lb-agent-chk then an agent
check may also be run using the agent-check parameter to a server,
which sets the TCP port to be used for the agent check.

e.g.
server	web1_1 127.0.0.1:80 check agent-port 10000

The agent-inter parameter may also be used to specify the interval
and timeout for agent checks.

If either the health or agent check determines that a server is down
then it is marked as being down, otherwise it is marked as being up.

Signed-off-by: Simon Horman <horms@verge.net.au>

---

v4
*  Increment global.maxsock for agent-port.

   If agent-port is configured then an extra socket is required.

* Do not send requests to secondary agent checks

  The request configuration of a proxy relates to the primary health
  check and should not be sent to the secondary health check if it
  is in operation.

* Correct usage of PR_O2_LB_AGENT_CHK

  The correct way to check for PR_O2_LB_AGENT_CHK is
  not to use x & PR_O2_LB_AGENT_CHK, but rather to use
  (x & PR_O2_CHK_ANY ) == PR_O2_LB_AGENT_CHK.

v2 - v3
* No change