MEDIUM: Add state to struct check
Add state to struct check. This is currently used to store one bit,
CHK_RUNNING, which is set if a check is running and clear otherwise.
This bit was previously SRV_CHK_RUNNING of the state element of struct
server.

This is in preparation for associating a agent check
with a server which runs as well as the server's existing check.

Signed-off-by: Simon Horman <horms+renesas@verge.net.au>

---

v8
* Use CHK_STATE_RUNNING instead of CHK_RUNNING

Notes by Willy on v6:
    Please use CHK_STATE_RUNNING instead of CHK_RUNNING so that it's
    easier to know it is related to the state attribute of the check.

    Also, we have many HCHK_* and now CHK_* as well, so I wonder how/when
    this will become confusing.