MEDIUM: Do not mark a server as down if the agent is unavailable
In the case where agent-port is used and the agent
check is a secondary check to not mark a server as down
if the agent becomes unavailable.

In this configuration the agent should only cause a server to be marked
as down if the agent returns "fail", "stopped" or "down".

Signed-off-by: Simon Horman <horms@verge.net.au>

---

v4
* Never update health of agent on failure to connect

  Previously the health was recharged on such cases.
  But this causes the health to be recharged even
  if the most recent result from the agent was "fail", "stopped", or "down".

  The result was unnecessary logging of "fail", "stopped", or "down"
  results from the agent if they were interleaved with timeouts.
  Or in other words, if the agent responded only occasionally
  but consistently with "fail", "stopped", or "down".

v3
* No change

v2
* Only log agent server status if the change in status
  will affect the server's state