MEDIUM: Add enable and disable agent unix socket commands
The syntax of this new commands are:

enable agent <backend>/<server>
disable agent <backend>/<server>

These commands allow temporarily stopping and subsequently
re-starting an auxiliary agent check. The effect of this is as follows:

New checks are only initialised when the agent is in the enabled. Thus,
disable agent will prevent any new agent checks from begin initiated until
the agent re-enabled using enable agent.

When an agent is disabled the processing of an auxiliary agent check that
was initiated while the agent was set as enabled is as follows: All
results that would alter the weight, specifically "drain" or a weight
returned by the agent, are ignored. The processing of agent check is
otherwise unchanged.

The motivation for this feature is to allow the weight changing effects
of the agent checks to be paused to allow the weight of a server to be
configured using set weight without being overridden by the agent.

Signed-off-by: Simon Horman <horms@verge.net.au>

---
v7
* Manual rebase
* Use enable|disable agent <backend>/<server> instead of
  set agent <backend>/<server> pause|unpause.
  - This is a reasonably trivial re-write of most of the patch
* Rename patch from "MEDIUM: Add set agent pause|unpause unix socket command"
  to "MEDIUM: Add enable and disable agent unix socket commands"

* Use CHK_STATE_DISABLED instead of CHK_DISABLED
  This is consistent with CHK_STATE_RUNNING which was
  previously named CHK_RUNNING
* Correct typos in comment
* Consistently don't call server_parse_weight_change_request() if stopped

Status: pending

Notes by Willy on v6 of
"MEDIUM: Add set agent pause|unpause unix socket command":
    I think that instead of setting the agent status to pause/unpaused, we'd
    rather reuse the same principle as we currently have with servers :

         enable agent <backend>/<server>
         disable agent <backend>/<server>

    OK for the rest of course.

v6
* First post