MEDIUM: Never update health of agent on failure to connect
Previously the health was recharged on such cases.
But this causes the health to be recharged even
if the most recent result from the agent was "fail", "stopped", or "down".

The result was unnecessary logging of "fail", "stopped", or "down"
results from the agent if they were interleaved with timeouts.
Or in other words, if the agent responded only occasionally
but consistently with "fail", "stopped", or "down".

Signed-off-by: Simon Horman <horms@verge.net.au>