MEDIUM: Set rise and fall of agent checks to 1
This is achieved by moving rise and fall from struct server to struct check.

After this move the behaviour of the primary check, server->check is
unchanged. However, the secondary agent check, server->agent now has
independent rise and fall values each of which are set to 1.

The result is that receiving "fail", "stopped" or "down" just once from the
agent will mark the server as down. And receiving a weight just once will
allow the server to be marked up if its primary check is in good health.

This opens up the scope to allow the rise and fall values of the agent
check to be configurable, however this has not been implemented at this
stage.

Signed-off-by: Simon Horman <horms@verge.net.au>

---

v7
* Manual Rebase
* Remove residual instance of 'server->' in comment in struct check hunk

Notes by Willy on v6:
    One remaining "server" below :

    -       int health;                             /* 0 to server->rise-1 = bad
+;
    -                                                * rise to server->rise+serv
+er->fall-1 = good */
    +       int health;                             /* 0 to rise-1 = bad;
    +                                                * rise to rise+server->fall
+-1 = good */
    +       int rise, fall;                         /* time in iterations */

    Otherwise OK.

v6
When checking health against rise in set_server_up the health of a check
should be checked against the rise of the same check. Otherwise the condition
for the s->agent may be false when it should be true as s->agent's rise and
thus healthy health value is 1 whereas the default rise for check is likely to
be 3, as check may be s->check and the the default value for s->check.rise is 3.

v5
* Rebase