MEDIUM: stats: Add JSON output option to show (info|stat)
Add a json parameter to show (info|stat) which will output information
in JSON format. A follow-up patch will add a JSON schema which describes
the format of the JSON output of these commands.

The JSON output is without any extra whitespace in order to reduce the
volume of output. For human consumption passing the output through a
pretty printer may be helpful.

e.g.:
$ echo "show info json" | socat /var/run/haproxy.stat stdio | \
     python -m json.tool

STAT_STARTED has bee added in order to track if show output has begun or
not. This is used in order to allow the JSON output routines to only insert
a "," between elements when needed. I would value any feedback on how this
might be done better.

Signed-off-by: Simon Horman <horms@verge.net.au>
---

For the simple configuration below a comparison of the size of info
and stats output is as follows:

$ show stat       =>                      1654 bytes
$ show stat typed =>                      7081 bytes
$ show stat json  =>                     45331 bytes
$ show stat json (pretty printed[*]) => 113390 bytes

$ show info       =>                      527 bytes
$ show info typed =>                      937 bytes
$ show info json  =>                     5330 bytes
$ show info json (pretty printed[*]) => 11456 bytes

[*] pretty printed using python -m json.tool

--- begin config ---
global
	daemon
	stats socket /tmp/haproxy.stat mode 600 level admin
	pidfile /tmp/haproxy.pid
	log /dev/log local4
	tune.bufsize 16384
	tune.maxrewrite 1024

defaults
	mode http
	balance roundrobin
	timeout connect 4000
	timeout client 42000
	timeout server 43000
	log global

listen VIP_Name
	bind 127.0.0.1:10080 transparent
	mode http
	balance leastconn
	cookie SERVERID insert nocache indirect
	server backup 127.0.0.1:9081 backup  non-stick
	option http-keep-alive
	option forwardfor
	option redispatch
	option abortonclose
	maxconn 40000
	log global
	option httplog
	option log-health-checks
	server RIP_Name 127.0.0.1  weight 100  cookie RIP_Name agent-check agent-port 12345 agent-inter 2000 check port 80 inter 2000 rise 2 fall 3  minconn 0 maxconn 0s on-marked-down shutdown-sessions disabled
	server RIP_Name 127.0.0.1  weight 100  cookie RIP_Name agent-check agent-port 12345 agent-inter 2000 check port 80 inter 2000 rise 2 fall 3  minconn 0 maxconn 0s on-marked-down shutdown-sessions
--- end config ---

Changes since RFC:
* Handle cases where output exceeds available buffer space
* Document that consideration should be given to updating
  dump functions if struct field is updated
* Limit JSON integer values to the range [-(2**53)+1, (2**53)-1] as per
  the recommendation for interoperable integers in section 6 of RFC 7159.